pyproject.toml
[project]
name = "task-manager-app-1"
version = "2.3.1"
description = "POETRY"
authors = [
    {name = "jocyeperelle",email = "perellepouokam@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.1"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
